requirements.txt
flask
python-dotenv
pandas
flask-cors